LE_Kotlin: extracted remaining abstractions from GameSpace.kt & moved these into "geometry" package
