+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/Border.kt
@@ -1,10 +1,6 @@
-package elements
-
-sealed interface GameSpace
+package geometry
 
 /**
  * a spot with any coordinate which does not exist on the reasonably sized field - is considered to be its border
  */
 internal data object Border : GameSpace
-
-data class OneAxis(val l: Int)
