+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/GameSpace.kt
@@ -1,0 +1,3 @@
+package geometry
+
+sealed interface GameSpace
