+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/OneAxis.kt
@@ -1,0 +1,3 @@
+package geometry
+
+data class OneAxis(val l: Int)
