+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/OneMoveProcessing.kt
@@ -1,7 +1,5 @@
 package geometry
 
-import elements.Coordinates
-
 interface OneMoveProcessing {
 
     fun getMaxLengthAchievedForThisMove(where: Coordinates): Int?
