+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept2D/Coordinates2D.kt
@@ -1,7 +1,7 @@
 package geometry.concept2D
 
-import elements.*
-import elements.Border
+import geometry.*
+import geometry.Border
 import geometry.LineDirectionFor1Axis
 
 // extensible alternative for CoordinatesXY
