+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept3D/Coordinates3D.kt
@@ -1,7 +1,7 @@
 package geometry.concept3D
 
-import elements.*
-import elements.Border
+import geometry.*
+import geometry.Border
 import geometry.LineDirectionFor1Axis
 
 // 3d as simple as 2d :)
