+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/concept3D/NearestAreaScanWith3D.kt
@@ -1,6 +1,6 @@
 package geometry.concept3D
 
-import elements.Coordinates
+import geometry.Coordinates
 import logic.GameField
 import geometry.OneMoveProcessing
 import logic.PlayerProvider
