+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/conceptXY/NearestAreaScanWithXY.kt
@@ -1,6 +1,6 @@
 package geometry.conceptXY
 
-import elements.Coordinates
+import geometry.Coordinates
 import logic.GameField
 import geometry.OneMoveProcessing
 import logic.PlayerProvider
