+++ Attt_LocalEngine_Kotlin/src/main/kotlin/logic/GameField.kt
@@ -5,7 +5,7 @@ import constants.MIN_GAME_FIELD_SIDE_SIZE
 import constants.SYMBOL_FOR_ABSENT_MARK
 import constants.SYMBOL_FOR_DIVIDER
 import constants.SYMBOL_FOR_NEW_LINE
-import elements.*
+import geometry.Coordinates
 import geometry.OneMoveProcessing
 import geometry.conceptXY.NearestAreaScanWithXY
 import publicApi.AtttPlayer
